More UI animations, begin Scenario conversion
A good portion of the battle UI uses Scenario for animation now, that
turned out to be an awesome tool!  The Tween class will be redundant
soon, Scenario makes it much easier to coordinate animations.  Also, the
starting party level is customizable instead of having it hard-coded
into the Party class.