Implement targeting menu, turn preview
Yay! Now you can select which battler gets hit by an attack!  While
targeting, the target's full name, status effects and HP/MP are shown
below the turn preview. As for the turn preview itself, that's
implemented. It updates in realtime while choosing a move or item... or
at least it would, if the turn predictor hadn't turned out to be
hopelessly broken!  I guess that's the next thing on the agenda...