Fix turn predictor
The turn predictor wasn't really broken, it was just a dumb off-by-one
error causing it to misinterpret the battler's next actions.  Also
changed the targeting cursor to show the name of the move being used
instead of targeted battler's name (the full name is still displayed in
the info box and seeing the move being targeted is very helpful). Seems
to be more accurate now. :-)  Getting very close to Tech Demo material
here...