Rebalance stat value calculation
When I gave myself a bit of time to think about the variables involved
in the damage calculation and how they would interact under different
operations, I realized the reason I couldn't properly balance damage
under the existing system was simply due to the sheer amount of variance
in calculated stats--so I rebalanced it. Hopefully now this will be the
last time I have to rewrite the damage formula!