Fix for Scenario slowdown
Stupid me, after the Scenario overhaul I forgot to remove the explicit
synchronize command in the run() method, originally meant to prevent the
camera and such from being reset until the rest of the scene was done.
Every time Scenario.run() was called, another synchronize command was
added to the end of the queue; after the scene is run enough times, this
causes massive delays while Scenario executes a series of useless
synchs.