Improve status overruling behavior
If you're under the effects of a status that overrules another and you
get hit with that other status (for example if you have Skeleton and an
enemy uses Necromancy on you), the overruling still applies and the new
status won't take effect.  Not too thrilled with the way this is
implemented, but it works.