Fix Rank 0 bug
Intuitively, a Rank 0 move would have no recovery time, i.e. you get
your next turn immediately.  The problem: If a unit's counter was set to
zero (as would happen in the case of a Rank 0 attack), said unit would
never get another turn due to the counter going negative on the next
tick.  This was a simple one-line fix.