Allow hiding speaker in textboxes
The opening narration isn't nearly as effective emotionally when Bruce's
name is displayed constantly, but simply passing null for a speaker
causes the full width of the textbox to be used for dialogue, which
hurts it aesthetically.  So now the talk scenelet takes an additional
argument, 'showSpeaker', which if set to false, hides the speaker's name
but aligns the text as if the name were visible.