Add status event 'dying', fix Zombie-Skeleton
Death-suppression is more flexible as it is now handled by a separate
status event instead of being rolled in with damage handling.
Zombie-Skeleton progression was nearly impossible to implement with the
old system.  This also fixes the long-standing issue of priority damage
still being able to kill a unit whose death should be suppressed.