RSB II: Shift phase transitions, nerf Phase 5
Robert's phases now change at lifebar boundaries, and Phase 5 had its
difficulty drastically reduced, since it's meant to be mostly symbolic
of Robert's desperation--the real fight is over.

Alcohol now recovers MP in addition to HP, and the full heal from said
item is ignored in the case of Zombie, closing the book on a very
long-lived cheat.  On the downside, Drunk now starts with a 50% AGI
drop, up from 25%.  Also updated the Immune status not to neutralize
Counterattack: this caused Immunized units trying to use Counter Stance
to never get another turn.