New damage formulas, update damage display
The damage formulas seem much more sane now, however this had the
unfortunate side-effect of lowering overall damage output in the RSB
fight.  Time will tell if this is for better or worse.  There is also a
new effect when displaying damage in-battle: The digits in the value
drop separately, similar to Final Fantasy.  This makes it more visually
striking when an attack does triple-digit damage as opposed to
double-digit.