RSB II: Fix Omni lock-in in Phase 3
If afflicted with Ignite or Frostbite three or more times in Phase 3,
Robert will retaliate with Omni after throwing up a Vaccine, if he has
any.  If he was afflicted once more after this, he wouldn't be able to
use the Vaccine (since he only has one) and therefore became locked into
Omni for the rest of the phase.