RSB II: Fix lack of Quickstrike combos in Phase 3
Due to a programming error, Robert was never initiating a Quickstrike
combo in Phase 3 even when the conditions to do so were otherwise met.
This meant that he used nothing but magic in P3--afflicting him with
Crackdown at this point completely shut him down.

I also took the opportunity to make Robert's P3 counterstrategies for
Skeleton and Crackdown much more subtle.