RSB II strategy updates, miscellaneous tweaks
There was a chance Robert could be tricked into attacking himself in
Phase 2 even after Frostbite or Ignite were removed for him.  This fixes
that as well as adding a Holy Water use at the start of Phase 4 if
needed, to try to set up for Alcohol.

This also lowers the action rank for item use and updates counterattacks
to have their power boosted in proportion to the attack being
countered--it's no longer possible to counter Quickstrike for double
damage.