Redesign move selection menu
The new design makes the rank of a move more visible.  This is
important, as the rank of a move is one of the most important factors in
its strategic value.  The MP cost of a move is secondary, at best, and
thus is now displayed to the right of the move's name.