Another Munch tweak, attack menu polishing
The Rank indicators in the attack menu are now dimmed for disabled
items, making it clearer at a glance that a move is unusable.

Accuracy for Munch is no longer dependent on user's current health, only
the target's, and target tier is taken into account as well.  This will
make it relatively easy to pull off against normal enemies, but
decidedly more difficult for bosses, as it should be.