Misc. rebalancing, RSB II third phase tweak
Robert now uses Electrocute in Phase 3 to zombify Scott before trying to
Power Tonic him to death.  Necromancy was too cheap a tactic in this
case, since the player is conditioned in earlier phases to delay curing
it.  Electrocute makes for a more obvious tell, and also serves as a
sneak peak of the move before Phase 4, when it becomes common.