Update Link to v0.2.14, other tweaks
Link 0.2.14 fixes the bug in invoke which caused the Stance regression
several commits ago.

Also added an optional 'addStatus' property to damage effects, to avoid
the necessity of an extra addStatus effect on the move; a statusChance
can also be added to the effect to indicate a less-than-100% chance of
inflicting the status.

This also further nerfs Counter Stance, which now doesn't reduce damage
at all, to compensate for the guaranteed 1.5x bonus.