Fix Spectral Kick glitch
The first time the Headless Horse used Spectral Kick, it missed, even
with the prioritized self-targeted effects.  This turned out to be
because the cached status info for the unit wasn't refreshed when it
took on a new status, so the Horse's new Ghost status wasn't being taken
into account during the accuracy check.