RSB II: Make Phase 5 more interesting
Phase 5 as originally conceived was incredibly dull, which is really not
a good thing when it's preceded by a full heal and takes some effort to
prevent it from happening.  So now he uses a ton of magic, with the
average power of his attacks going up as he gets closer to defeat.

Also added some use of Charge Slash by Robert in Phase 2, as that phase
was going by much too quickly, making Phase 3 feel like hitting a brick
wall at 100 mph.