Item prediction for AIs, improved RSB II strategy
I'm realizing as I work on Robert's AI that determinism isn't a bad
thing--it's what allows the player to make strategies in the first
place.  The trick is to predicate the clauses on things that are
difficult to control, such as turn order, to ensure that the boss's
response is hard to predict.