Yet more work on Headless Horse, RSB II
It's much harder to design a good AI when there are 3 battlers on the
other side.  As for Robert, if the player tricks him into using his last
Holy Water despite the conservation clause in Phase 2 (namely, by
inflicting Zombie and feeding him healing items), he will forgo using
his Vaccine in P3 in favor of saving it to prepare for the full heal and
final stretch.  Additionally, Omni is no longer blockable via Guard--it
does the full damage regardless.  This makes an Omni reprisal from
Spellbinder much, much harder to exploit for free turns.