RSB II: Overhaul Zombie exploitation strategy
This now works on an "alert level" system, with the alert level rising
each time Zombie is exploited to do tons of damage to Robert with
healing items.  Note that the alert level also rises on its own as phase
transitions happen--exploitation in P4 will draw an immediate Omni, for
example.