Major RSB II overhaul, minor update to Stances
Robert is a bit more predictable now, giving the player more ability to
see painful attacks such as Upheaval coming ahead of time.  This
philosophy, which will be considered for all bosses going forward,
elevates Guard from a useful auxiliary ability to an integral part of
the battle system.  As for Guard and Counter, they are now broken
prematurely if an attack misses.