RSB II: P4/P5 edits, edit to Scott's speech
Phase 4 is a separate phase now, instead of just being an excuse to prep
for Phase 5.  Phase 5 now starts before Alcohol, with #9's Desperation.

For the Alcohol intro, Scott's dialogue has been improved and Robert's
lines changed to silence.  Robert is desperate to end the fight at that
point, but exhausted--not exactly the right state of mind for bluster.