Various RSB II tweaks, revert to old stat curve
Admittedly I hadn't thought the change to the stat curve through.  A
dynamic range of 4 meant that a unit with, say, 0 base DEF would take
the equivalent of 4x damage from an attacker with 100 base STR.  On that
note, I also nerfed the damage increase from Off Guard; the 2x weakness
during the first turn of Charge Slash is just too steep a price to pay.