Revamp field conditions, add more events
Field conditions are now notified when a unit is afflicted or cured of a
status, as well as when a unit is healed.  Also, if an effect, such as
damage, is suppressed by a field condition, the corresponding status
event will not be invoked at all.  Additionally, several field
conditions were reimagined to affect a random battler each cycle instead
of all of them at once.  This keeps things more balanced.

This also splits Scott's perpetual Spectacles healing into its own
status (Specs Aura): ReGen is now time-limited but restores more health
at a time, which better fits the spirit of Rejuvenate.  On that note,
this also updates healing spells to take the user's FOC into account,
just as damaging magic does.