Fix crash bug in Scott Temple AI, other glitches
Off Guard no longer affects residual damage, Subzero damage is
calculated correctly, and the Temple AI no longer crashes the engine
when responding to player healing.  Plus a cosmetic update--the status
prevention indicator ("immune"/"ward") is only shown if the status is a
move's primary effect.  If the status infliction is secondary (for
example, Zombie for Electrocute), no indicator is shown, to prevent
information overload.