Yet more work on Temple, Inferno/Subzero tweaks
Inferno and Subzero, in addition to boosting same-element attacks, now
also weaken the opposite element.  This adds another element of strategy
to the Temple fight: You can remove the condition he installs in Phase 1
instead of waiting for him to do it at the start of P2, but you would
take more damage from the follow-up attack as a result.