Start work on RSB I, add Lauren, enhancements
Added Lauren as a battler, along with corresponding shuriken (throwing
star) skills.

Making use of the new RNG library, phase-transition thresholds for
bosses are now on a bell curve, as opposed to the uniform variance they
had before.  This keeps the average amount of variance about the same
while increasing the overall range, making it more difficult to predict
a phase change.

Also, an overhaul to the Scott Starcross AI to make it less fragile;
combo selection will now work properly regardless of which (and how
many) units are in play, and he will scale back if one of his opponents
is killed, rather than redirecting everything to the remaining unit(s).