Add new Pact object to minipromises
minipromises... seems awkward.  Maybe I should rename it to minipact?

In any case, Pacts vastly simplify promise management for libraries such
as minithreads that hand out a lot of promises.  A given Pact object can
only be used to resolve Promises made from it, which avoids the issues
inherent to adding public .resolve() and .reject() methods to promises.
The "pactmaker", as it were, holds onto the Pact and is then able to
safely hand out the immutable Promise objects as usual.

Without something like this, the alternative is to create a lot of
one-off callbacks to call back into the promise to fulfill it.  Not only
is this inefficient, but it makes code much harder to follow!