Fix long-standing attack menu crash
This was a ridiculously difficult bug to reproduce.  Thanks to the
Sphere Studio debugger (SSJ wasn't yet mature enough for this one, it
seems), I finally tracked down the cause.  The fix: A one-line change.
Go figure.

Here's what happened: If you opened a menu drawer with a bunch of items
and highlighted one towards the bottom, then closed the drawer and moved
to a one with fewer items, the game would crash--but _only_ if you
pressed the Cancel button *before* the Confirm button.  That's why it
seemed so random and rarely turned up in practice.