Use Sphere v1 compatibility mode
Updating the Specs Engine to run in native Sphere v2 mode is going to be
a huge undertaking that I'm not ready to take on just yet.

On the plus side, all the fancy Sphere v2 APIs are still available in
compatibility mode--I just don't get the benefit of everything being a
CommonJS module.