Update minimum miniSphere version
The README claimed minisphere 3.3 as the minimum requirement, which...
isn't exactly true anymore.  To avoid that in the future since
miniSphere is under active development, I removed any reference to a
required version and just pointed to the release page.

I also updated the build instructions.  Cell no longer requires any
command line options to do a basic build, which is nice.