Experiment with rebalancing HP/damage
The original battle system design called for bosses to take less damage
than standard enemies.  This was meant to illustrate how outclassed the
player characters are, but in actual gameplay, it ends up being
frustrating since there's no way for the player to close the gap.  It's
not too nice to take 800+ damage from a boss and only ever be able to do
200ish in return.