Use dollar-path to load inGameClock module
The dollar-sign prefix is something miniSphere should have had a while
ago.  I think I had an automatic aversion to it because it kind of
dimininishes the "purity" of SphereFS, but pragmatically, the main
script is tied with the game manifest for most important file in the
whole game package (the game can't function without it), so it makes
sense that it has a prefix.  Plus it makes module imports more readable
(and maintainable) given how the vast majority of people organize source
code.