Refactor codebase, make more modular
I'm technically not done refactoring yet, but this is the lion's share
of the work.  I split apart several modules that were trying to do too
much, renamed everything back to camelCase (as it was before), and use
default exports wherever possible.  I reserve the right to go back on
that last one...