Update menus to use v2 Font and Color
The number of temporary Color objects I create and discard every frame
is insane.  All the battle menus really need to be refactored; it'll be
easier if they're already using the Sphere v2 Color and Font APIs.

Prim usage is way out of hand as well.  Alas, a project for another
day...