Enable TypeScript support for Specs
Using the TS build script in a largish project like this will help me
ensure more edge cases are covered.  I ended up missing a lot of bugs
when I was just using a small test project for it.

Note that Specs now requires up-to-date (post-5.3) builds of both
miniSphere and Cell starting with this commit.  I hate doing that,
but...