Lower POW ratings of magic
I realized that the reason I've been having so much difficulty balancing
damage (or, at least, one of the reasons) is that there's too much
dynamic range on attack power ratings. I've been scaling based on Omni,
but I forgot that Omni is meant to be exceptional in its damage output,
so regular, everyday attacks probably shouldn't exceed half of its
output. Especially given that the baseline for normal attacks is
15POW...